Implementation of Wikipedia perceptron
This was my first cut at implementing a perceptron. This was based on
the wikipedia description of a perceptron, which differs in some ways
from that of the course. It was useful in helping me get the scaffolding
in place for generating and testing vectors, but it's incomplete and
doesn't work very well.